Removes useless code
